feat: session-state and auto-app rerun example
Streamlit reloads the page each time the user interacts with the input elements, and does this from
top to bottom. Hence, this means that we can also re-use some of the input provided. To do so, we
can collect the value from the `session_state`. Note that the `session_state` will be empty at the
start of a new browser session.