feat: add timeslot filtering for the programme
For filtering the schedule, we use a multiselect filter. Before displaying the result, we should make
sure that we apply the filter. Instead of using the return value of the multiselect box, we will
rely on the use of the session state. This makes the app in general easier to understand + maintain.