feat: select from (pre-filtered) overview of streams
Adding a second layer of filters with filtering by streams. A stream is to some extent in an hiererachical
way connect to the timeslot filter: in some timeslots some streams will not be organized, though in
the majority of cases they are. Having dynamically connected filters in streamlit is not easy and
requires customization or using customized components. Hence, we take a middle ground solution here
with the timeslots influencing which streams are shown