feat: make abstracts in filtered results selectable
From the filtered set of options, users can select the abstracts they would which to look a bit
deeper into. In order to do so, we make the dataframe selectable and capture the events
originating from the user interacting with the dataframe.

Note that if the filters on top of the table change, the selection in the programme gets
reset.