feat: model to optimize session attendance utility
During a conference, we want to get the most out of attending different sessions. Hence,
we will integrate an optimization model which maximizes the total utility we get from
attending conference sessions. Note that we assume we do not want to swap rooms during
a session, and the mean utility of a session is the combined value of all talks included
in it.

The decision then comes down to select the best possible set of sessions out of the
available one such that we have at most one session per timeslot.